+++ test/cherry/compiler_test.cljs
@@ -5,6 +5,7 @@
    [cherry.embed-test]
    [cherry.html-test]
    [cherry.jsx-test]
+   [cherry.macro-test]
    [cherry.squint-and-cherry-test]
    [cherry.test-utils :refer [js! jss! jsv!]]
    [clojure.string :as str]
@@ -531,4 +532,4 @@ IReset (-reset! [this v]
 
 (defn init []
   (cljs.test/run-tests 'cherry.compiler-test 'cherry.jsx-test 'cherry.squint-and-cherry-test
-                       'cherry.html-test 'cherry.embed-test))
+                       'cherry.html-test 'cherry.embed-test 'cherry.macro-test))